fix: remove GestureHandler.web.js and instead reuse GestureHandler.js
