--- AnimatedEvent.js
@@ -1,3 +1,0 @@
-import { Animated } from 'react-native';
-
-export default Animated.Event;
