+++ GestureHandlerButton.native.js
@@ -1,0 +1,8 @@
+import { requireNativeComponent } from 'react-native';
+
+const RNGestureHandlerButton = requireNativeComponent(
+  'RNGestureHandlerButton',
+  null
+);
+
+export default RNGestureHandlerButton;
