+++ PlatformConstants.native.js
@@ -1,0 +1,3 @@
+import { NativeModules } from 'react-native';
+
+export default NativeModules.PlatformConstants;
