+++ Swipeable.js
@@ -6,7 +6,6 @@
 
 import React, { Component } from 'react';
 import { Animated, StyleSheet, View } from 'react-native';
-import { AnimatedEvent } from './AnimatedEvent';
 
 import { PanGestureHandler, TapGestureHandler, State } from './GestureHandler';
 
@@ -59,7 +58,7 @@ export default class Swipeable extends Component<PropType, StateType> {
     friction: 1,
     useNativeAnimations: true,
   };
-  _onGestureEvent: ?AnimatedEvent;
+  _onGestureEvent: ?Animated.Event;
   _transX: ?Animated.Interpolation;
   _showLeftAction: ?Animated.Interpolation | ?Animated.Value;
   _leftActionTranslate: ?Animated.Interpolation;
